# plugwerk-api/src/main/resources/openapi/plugwerk-api.yaml
openapi: 3.1.0
info:
  title: Plugwerk API
  description: |
    **Plugwerk** is a self-hosted plugin marketplace for the [PF4J](https://pf4j.org/) ecosystem.
    It lets teams publish, version, and distribute Java/Kotlin plugins to their own applications
    without relying on a public registry.

    ## Core Concepts

    ### Namespaces
    A **namespace** is the top-level organisational unit. Every plugin belongs to exactly one
    namespace. Namespaces are identified by a URL-safe **slug** (lowercase alphanumeric + hyphens,
    2–64 characters). You might use one namespace per product, team, or customer, e.g. `acme-core`.

    ### Plugins
    A **plugin** is a logical grouping of releases for a single PF4J plugin ID. The `pluginId`
    matches the `Plugin-Id` entry in the PF4J manifest (`MANIFEST.MF` or `plugin.properties`).
    Each plugin can have a human-readable name, description, icon, and categorisation metadata.

    ### Releases
    A **release** is a specific versioned artifact (JAR or ZIP) for a plugin. Versions follow
    [Semantic Versioning](https://semver.org/) (`MAJOR.MINOR.PATCH`). Releases go through a
    lifecycle: `draft` → `published` → `deprecated` / `yanked`. Only `published` releases are
    returned to catalog and update-check consumers. A namespace owner can optionally require
    manual review before a release is published (see the **Reviews** tag).

    ### Plugin Descriptor (`plugwerk.yml`)
...
    ```yaml
    plugwerk:
...
      version: 1.2.0
...
    ```
...
servers:
...
security: []
...
paths:
...
  /auth/change-password:
    post:
      tags:
        - Auth
      summary: Change own password
      description: |
        Changes the authenticated user's password. Requires the current password for verification.

        After a successful change, `passwordChangeRequired` is set to `false` on the user account.
        This endpoint must be called before accessing any other resource when the login response
        contains `passwordChangeRequired: true`.
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/users:
    get:
      tags:
        - AdminUsers
      summary: List all users
      operationId: listUsers
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of all local users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - AdminUsers
      summary: Create a new user
      operationId: createUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created. The user must change the password on first login.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/users/{userId}:
    patch:
      tags:
        - AdminUsers
      summary: Update user (enable/disable or reset password)
      operationId: updateUser
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - AdminUsers
      summary: Delete user (superadmin cannot be deleted)
      operationId: deleteUser
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Cannot delete the superadmin account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      description: Standard error response body returned for all 4xx and 5xx responses
      properties:
        status:
          type: integer
          description: HTTP status code (mirrors the response status)
          example: 404
        error:
          type: string
          description: Short machine-readable error code (e.g. `NOT_FOUND`, `VALIDATION_ERROR`)
          example: NOT_FOUND
        message:
          type: string
          description: Human-readable description of what went wrong
          example: Plugin 'com.example.my-plugin' not found in namespace 'acme-core'
        timestamp:
          type: string
          format: date-time
          description: ISO-8601 timestamp of when the error occurred
      required:
        - status
        - error
        - message
        - timestamp
...
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
      required:
        - currentPassword
        - newPassword
    UserDto:
      type: object
      description: Local Plugwerk user account
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          nullable: true
        enabled:
          type: boolean
        passwordChangeRequired:
          type: boolean
        isSuperadmin:
          type: boolean
          description: When true the user has global superadmin privileges and cannot be deleted
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - enabled
        - passwordChangeRequired
        - isSuperadmin
        - createdAt
    UserCreateRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 255
        email:
          type: string
          format: email
          nullable: true
        password:
          type: string
          minLength: 8
      required:
        - username
        - password
    UserUpdateRequest:
      type: object
      description: Partial update — include only the fields to change
      properties:
        enabled:
          type: boolean
          nullable: true
        newPassword:
          type: string
          minLength: 8
          nullable: true
...
  responses:
    BadRequest:
      description: The request body or parameters failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication credentials are missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived JWT token obtained via `POST /api/v1/auth/login`.
        Include as `Authorization: Bearer <token>`. Default validity: 8 hours.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        Long-lived API key for CI/CD pipelines and service accounts.
        Include as `X-Api-Key: <your-api-key>` header.
        API keys are provisioned by the server administrator.